renamed Excpetions to Exceptions - and updated class imports
