+++ src/Struct.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Spatie\Typed;
 
 use ArrayAccess;
-use Spatie\Typed\Excpetions\WrongType;
-use Spatie\Typed\Excpetions\UninitialisedError;
+use Spatie\Typed\Exceptions\WrongType;
+use Spatie\Typed\Exceptions\UninitialisedError;
 
 class Struct implements ArrayAccess
 {
