+++ tests/TypeTest.php
@@ -18,7 +18,7 @@ use Spatie\Typed\Types\GenericType;
 use Spatie\Typed\Types\IntegerType;
 use Spatie\Typed\Types\CallableType;
 use Spatie\Typed\Types\CollectionType;
-use Spatie\Typed\Excpetions\InferredTypeError;
+use Spatie\Typed\Exceptions\InferredTypeError;
 
 class TypeTest extends TestCase
 {
